v0.30
keyboard control("W","A","S","D")

+++ VertexBuffer.h
@@ -1,5 +1,5 @@
 #pragma once
-#include<d3d11.h>
+#include <d3d11.h>
 
 class DeviceContext;
 
@@ -7,7 +7,7 @@ class VertexBuffer
 {
 public:
 	VertexBuffer();
-	bool load(void* list_vertices, UINT size_vertex, UINT size_list, void* shader_byte_code, UINT size_byte_shader);
+	bool load(void* list_vertices, UINT size_vertex, UINT size_list, void* shader_byte_code, size_t size_byte_shader);
 	UINT getSizeVertexList();
 	bool release();
 	~VertexBuffer();
@@ -20,4 +20,3 @@ private:
 private:
 	friend class DeviceContext;
 };
-
